refactor: Add map_to_gnss transform broadcasting
This commit adds the functionality to broadcast the transform from the "map" frame to the "map_to_gnss" frame. This transform is necessary for accurate positioning of the GNSS data in the map frame. The code changes include the addition of the `publishMapToGnssTransform` function and the creation of a `tf2_ros::TransformBroadcaster` object in the `RosPospacBridge` class constructor.