refactor: Update ros_pospac_bridge.cpp to include GNSS_INS pose array publishing
This commit updates the ros_pospac_bridge.cpp file to include the publishing of a pose array for GNSS_INS. The pose array is populated with the accumulated GNSS_INS poses and is published under the topic "/ros_pospac_bridge/gnss_ins_pose_array". This addition allows for the visualization of GNSS_INS poses on the PCD map.